Move all pre-path formatting routines to `._pathops`, proto formatter type
